requirements.txt
pandas>=1.3.0
numpy>=1.20.0
requests>=2.25.0
plotly>=5.0.0
dash>=2.0.0
scikit-learn>=1.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0
openpyxl>=3.0.0 

aiohttp>=3.8.0
python-dotenv>=0.19.0 